requirements.txt
moviepy
google-api-python-client
openai
gtts
python-dotenv
requests
shutilwhich
urllib3
pytube
opencv-python
